refactor: 좌표 두 개 스냅 기능을 Coordinate API로 분리
- 문서화 API에 AUTH 명시
